backend/requirements.txt
fastapi==0.111.0
orjson==3.10.3
uvicorn[standard]==0.30.1
uvloop==0.19.0;sys_platform!='win32'
httptools==0.6.1
pydantic==2.7.1
python-dateutil==2.9.0
psutil==5.9.8
pytest==8.2.2
httpx==0.27.0
pytest-asyncio==0.23.7
